requirements.txt
python-dotenv==1.0.1
Pillow==10.3.0
google-generativeai==0.5.4
streamlit==1.35.0